blogをコンポーネント化、不要ファイルとコード削除
