--- src/components/twitterList.ts
@@ -1,14 +1,0 @@
-export const twitterList = [
-  {
-    name: "lightsound/nexst-tailwind",
-    tweet: "📣 新サービス「Noway Form」をリリースしました！Noway Formは、Notionのデータベースをもとにフォームを作成できるサービスです。これまでGoogle FormsでやっていたことがNotionだけで完結します✌✨試しに使っていただけると幸いです😊https://www.noway-form.com/ja"
-  },
-  {
-    name: "lightsound/nexst-tailwind",
-    tweet: "📣 新サービス「Noway Form」をリリースしました！Noway Formは、Notionのデータベースをもとにフォームを作成できるサービスです。これまでGoogle FormsでやっていたことがNotionだけで完結します✌✨試しに使っていただけると幸いです😊https://www.noway-form.com/ja"
-  },
-  {
-    name: "lightsound/nexst-tailwind",
-    tweet: "📣 新サービス「Noway Form」をリリースしました！Noway Formは、Notionのデータベースをもとにフォームを作成できるサービスです。これまでGoogle FormsでやっていたことがNotionだけで完結します✌✨試しに使っていただけると幸いです😊https://www.noway-form.com/ja"
-  },
-]
