+++ src/pages/index.tsx
@@ -7,10 +7,7 @@ import { portfolio } from "@/components/portfolioList";
 import Github from "@/components/githubComponent";
 import Twitter from "@/components/twitterComponent";
 import { client } from "@/libs/cmsClient";
-import Link from "next/link";
-import dayjs from "dayjs";
-import { useMediaQuery } from "react-responsive";
-import { FC, useEffect, useState } from "react";
+import { FC } from "react";
 import LinkButton from "@/components/button";
 import useSWR from "swr";
 import { apolloClient } from "@/libs/apolloClient";
@@ -18,11 +15,11 @@ import { GithubProps, GithubType } from "@/types/github";
 import { query } from "@/libs/getGithubData";
 import { Blog, BlogData } from "@/types/blog";
 import { Tweets } from "@/types/tweet";
+import { BlogList } from "@/components/blogList";
 import {
   Container,
   createStyles,
   SimpleGrid,
-  Stack,
   Text,
   Title,
   Loader,
@@ -80,58 +77,7 @@ const fetcher = async (url: string) =>
 
 const Home: FC<Props> = (props) => {
   const { data, error } = useSWR<Tweets>("/api/twitter", fetcher);
-  const [isClient, setIsClient] = useState(false);
   const { classes } = useStyles();
-  const isMobile = useMediaQuery({
-    query: "(max-width: 400px)",
-  });
-  const isLaptop = useMediaQuery({
-    query: "(min-width: 401px)",
-  });
-
-  const blog = () => {
-    if (isClient) {
-      if (isMobile) {
-        return props.blogData.contents.slice(0, 4).map((list, index) => (
-          <Link key={index} href={`/blog/${list.id}`} passHref>
-            <a>
-              <Stack key={index} className="mb-6">
-                <Title order={2}>{list.title}</Title>
-                <Text
-                  dangerouslySetInnerHTML={{
-                    __html: list.body,
-                  }}
-                />
-                {dayjs(list.createdAt).format("YYYY年MM月DD日")}
-              </Stack>
-            </a>
-          </Link>
-        ));
-      } else if (isLaptop) {
-        return props.blogData.contents.slice(0, 5).map((list, index) => (
-          <Link key={index} href={`/blog/${list.id}`} passHref>
-            <a>
-              <Stack key={index} className="mb-6">
-                <Title order={2}>{list.title}</Title>
-                <Text
-                  dangerouslySetInnerHTML={{
-                    __html: list.body,
-                  }}
-                />
-                {dayjs(list.createdAt).format("YYYY年MM月DD日")}
-              </Stack>
-            </a>
-          </Link>
-        ));
-      }
-    } else {
-      return <></>;
-    }
-  };
-
-  useEffect(() => {
-    setIsClient(true);
-  }, []);
 
   return (
     <div className={styles.container}>
@@ -145,12 +91,12 @@ const Home: FC<Props> = (props) => {
         <Title order={1} className={classes.heading}>
           Blog
         </Title>
-        <div>{blog()}</div>
+        <BlogList blogData={props.blogData} />
         <LinkButton text="View All" href="/blog" />
       </Container>
       <PortfolioSection portfolioSection={portfolio} />
       <SimpleGrid cols={2} spacing="xs">
-        <Container className="">
+        <Container>
           <Title order={1} className={classes.heading}>
             GitHub
           </Title>
